backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Auth and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation and settings
pydantic==2.6.1
pydantic-settings==2.1.0

# PDF generation and processing
reportlab==4.1.0
PyPDF2==3.0.1
Pillow==10.2.0
pdf2image==1.17.0  # Requires poppler-utils system package

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1
httpx==0.26.0

